Added Renderer3D - Added basic axis + rotating tri

+++ TheOneEditor/Module.h
@@ -34,7 +34,7 @@ public:
 	}
 
 	// Called each loop iteration
-	virtual bool Update(float dt)
+	virtual bool Update(double dt)
 	{
 		return true;
 	}
